egovframe-simple-homepage-template 5.0.0

+++ src/main/java/egovframework/let/cop/bbs/service/impl/BBSAddedOptionsDAO.java
@@ -1,10 +1,9 @@
 package egovframework.let.cop.bbs.service.impl;
+import org.egovframe.rte.psl.dataaccess.EgovAbstractMapper;
+import org.springframework.stereotype.Repository;
+
 import egovframework.let.cop.bbs.service.BoardMaster;
 import egovframework.let.cop.bbs.service.BoardMasterVO;
-
-import org.egovframe.rte.psl.dataaccess.EgovAbstractMapper;
-
-import org.springframework.stereotype.Repository;
 
 /**
  * 2단계 기능 추가 (댓글관리, 만족도조사) 관리를 위한 데이터 접근 클래스
@@ -32,7 +31,7 @@ public class BBSAddedOptionsDAO extends EgovAbstractMapper {
      * @param BoardMaster
      */
     public int insertAddedOptionsInf(BoardMaster boardMaster) throws Exception {
-	return (int)insert("BBSAddedOptionsDAO.insertAddedOptionsInf", boardMaster);
+	return insert("BBSAddedOptionsDAO.insertAddedOptionsInf", boardMaster);
     }
     
     /**
